requirements.txt
Django~=3.2.6
djangorestframework==3.12.2
django-cors-headers==3.5.0
pika==1.1.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
django-allauth==0.44.0